Add additionnal repo for TBB
ROCm 6.2.0 rocrand is linked against libtbb.so.12 but the default repository lead to
the installation of the libtbb2 package. This lead to a missing dependency. The missing
dependency then cause a failure of the initialization of the GPU backend of TensorFlow.

This additionnal repository need to be added after json-c installation. If done before,
the version of json-c that will be installed will be incompatible with libfabric.
An alternative is to explicitly specify the json-c version to be installed
(libjson-c-devel-0.13).