Install AMD-SMI in Conda environment
Since PyTorch v2.4 AMD-SMI is used for GPU monitoring in PyTorch. As it's not in the
PYTHONPATH by default, we either need to add it, or install it in the Conda environment
so that it can be found by PyTorch. We choose to do the later as it was the method
applied to the VLLM container.