Fix RCCL segmentation fault
PyTorch comes with RCCL from ROCm 6.0.0 which cause a segmentation fault. The
following backtrace was collected:

  ncclTopoCompute(ncclTopoSystem*, ncclTopoGraph*) () from /opt/miniconda3/envs/pytorch/lib/python3.12/site-packages/torch/lib/librccl.so
  initTransportsRank(ncclComm*, ncclComm*) () from /opt/miniconda3/envs/pytorch/lib/python3.12/site-packages/torch/lib/librccl.so
  ncclCommInitRankFunc(ncclAsyncJob*) () from /opt/miniconda3/envs/pytorch/lib/python3.12/site-packages/torch/lib/librccl.so
  ncclAsyncJobMain(void*) () from /opt/miniconda3/envs/pytorch/lib/python3.12/site-packages/torch/lib/librccl.so
  start_thread () from /lib64/libpthread.so.0
  clone () from /lib64/libc.so.6

The RCCL library from /opt/rocm doesn't produce a segfault, probably because a fix has
been implemented between the releases of ROCm 6.0.0 and 6.1.3
(https://github.com/ROCm/rccl/pull/1153).

To solve the issue, we copy the library from /opt/rocm to the PyTorch installation.